Refactor KPI type naming and update related references
Renamed KpiType enum values from NUMERICAL/BINARY to QUANTITATIVE/QUALITATIVE and moved KpiType to a constants package. Updated all backend and frontend references, validation logic, and UI rendering to use the new names. Also fixed minor UI text and property names for consistency.
