+++ backend/kpi-tracker-models/src/main/java/org/pahappa/systems/kpiTracker/models/kpi/KPI.java
@@ -1,5 +1,6 @@
 package org.pahappa.systems.kpiTracker.models.kpi;
 
+import org.pahappa.systems.kpiTracker.constants.KpiType;
 import org.pahappa.systems.kpiTracker.models.goalMgt.Goal;
 import org.sers.webutils.model.BaseEntity;
 import org.sers.webutils.model.security.User;
@@ -29,7 +30,7 @@ public class KPI extends BaseEntity {
     @Column(name = "kpi_type", nullable = false)
     private KpiType kpiType;
 
-    // Fields for NUMERICAL KPI
+    // Fields for Quantitative KPI
     @Column(name = "start_value", precision = 19, scale = 4)
     private BigDecimal startValue;
 
@@ -42,7 +43,7 @@ public class KPI extends BaseEntity {
     @Column(name = "unit_of_measure")
     private String unitOfMeasure;
 
-    // Field for BINARY KPI
+    // Field for QUALITATIVE KPI
     @Column(name = "is_complete")
     private Boolean isComplete = false;
 
@@ -57,7 +58,7 @@ public class KPI extends BaseEntity {
         }
 
         switch (this.kpiType) {
-            case NUMERICAL:
+            case QUANTITATIVE:
                 if (startValue == null || targetValue == null || currentValue == null || targetValue.subtract(startValue).compareTo(BigDecimal.ZERO) == 0) {
                     return BigDecimal.ZERO;
                 }
@@ -75,7 +76,7 @@ public class KPI extends BaseEntity {
                 }
                 return progress.setScale(2, RoundingMode.HALF_UP);
 
-            case BINARY:
+            case QUALITATIVE:
                 return (isComplete != null && isComplete) ? new BigDecimal("100") : BigDecimal.ZERO;
 
             default:
