+++ backend/kpi-tracker-services/src/main/java/org/pahappa/systems/kpiTracker/core/services/impl/KpiServiceImpl.java
@@ -4,7 +4,7 @@ import com.googlecode.genericdao.search.Search;
 import org.pahappa.systems.kpiTracker.core.services.KpiService;
 import org.pahappa.systems.kpiTracker.models.goalMgt.Goal;
 import org.pahappa.systems.kpiTracker.models.kpi.KPI;
-import org.pahappa.systems.kpiTracker.models.kpi.KpiType;
+import org.pahappa.systems.kpiTracker.constants.KpiType;
 import org.pahappa.systems.kpiTracker.utils.Validate;
 import org.sers.webutils.model.RecordStatus;
 import org.sers.webutils.model.exception.OperationFailedException;
@@ -21,7 +21,7 @@ public class KpiServiceImpl extends GenericServiceImpl<KPI> implements KpiServic
     @Override
     public KPI saveInstance(KPI kpi) throws ValidationFailedException, OperationFailedException {
         Validate.notNull(kpi, "KPI details cannot be null");
-        Validate.notNull(kpi.getGoal(), "KPI must be attached to a goal");
+        //Validate.notNull(kpi.getGoal(), "KPI must be attached to a goal");
         Validate.hasText(kpi.getTitle(), "KPI title is required");
         Validate.notNull(kpi.getKpiType(), "KPI type is required");
 
@@ -29,9 +29,9 @@ public class KpiServiceImpl extends GenericServiceImpl<KPI> implements KpiServic
             throw new ValidationFailedException("A KPI with the same title already exists for this goal.");
         }
 
-        if (kpi.getKpiType() == KpiType.NUMERICAL) {
-            Validate.notNull(kpi.getStartValue(), "Start value is required for a numerical KPI");
-            Validate.notNull(kpi.getTargetValue(), "Target value is required for a numerical KPI");
+        if (kpi.getKpiType() == KpiType.QUANTITATIVE) {
+            Validate.notNull(kpi.getStartValue(), "Start value is required for a Quantitative KPI");
+            Validate.notNull(kpi.getTargetValue(), "Target value is required for a Quantitative KPI");
             Validate.isTrue(kpi.getTargetValue().compareTo(kpi.getStartValue()) > 0, "Target value must be greater than the start value.");
 
             // Initialize current value if not set
@@ -39,16 +39,16 @@ public class KpiServiceImpl extends GenericServiceImpl<KPI> implements KpiServic
                 kpi.setCurrentValue(kpi.getStartValue());
             }
 
-            // Clear binary fields to ensure data integrity
+            // Clear QUALITATIVE fields to ensure data integrity
             kpi.setIsComplete(null);
 
-        } else if (kpi.getKpiType() == KpiType.BINARY) {
-            // Initialize binary fields if not set
+        } else if (kpi.getKpiType() == KpiType.QUALITATIVE) {
+            // Initialize QUALITATIVE fields if not set
             if (kpi.isNew() && kpi.getIsComplete() == null) {
                 kpi.setIsComplete(false);
             }
 
-            // Clear numerical fields to ensure data integrity
+            // Clear Quantitative fields to ensure data integrity
             kpi.setStartValue(null);
             kpi.setTargetValue(null);
             kpi.setCurrentValue(null);
