+++ frontend/kpi-tracker-web/src/main/resources/hibernate.cfg.xml
@@ -56,7 +56,7 @@
 		<mapping class="org.pahappa.systems.kpiTracker.models.goalMgt.Goal"/>
 		<mapping class="org.pahappa.systems.kpiTracker.models.goalMgt.GoalDepartment"/>
 		<mapping class="org.pahappa.systems.kpiTracker.models.kpi.KPI"/>
-		<mapping class="org.pahappa.systems.kpiTracker.models.kpi.KpiType"/>
+		<mapping class="org.pahappa.systems.kpiTracker.constants.KpiType"/>
