+++ frontend/kpi-tracker-web/src/main/webapp/pages/kpis/KpisView.xhtml
@@ -15,17 +15,17 @@
                             <div class="p-col-12">
                                 <div class="card">
                                     <p:dataTable id="kpisTable" var="model" value="#{kpiView.dataModels}"
-                                                 widgetVar="teamsTable"
+                                                 widgetVar="kpisTable"
                                                  paginatorTemplate="{CurrentPageReport}  {FirstPageLink} {PreviousPageLink} {PageLinks} {NextPageLink} {LastPageLink}"
                                                  paginator="true" lazy="true" rows="20"
-                                                 emptyMessage="No teams found."
+                                                 emptyMessage="No kpis found."
                                                  paginatorPosition="bottom" paginatorAlwaysVisible="false"
                                                  rowIndexVar="row" reflow="true" styleClass="TableAlgnment">
 
                                         <f:facet name="header">
                                             <div class="p-d-flex p-jc-between">
                                                 <div>
-                                                    <span class="table-header-text">Key Perfomance Indicators</span>
+                                                    <span class="table-header-text">Key Performance Indicators</span>
                                                 </div>
                                                 <div>
                                                     <div class="p-d-flex p-flex-wrap">
@@ -50,19 +50,7 @@
                                             <f:facet name="header">
                                                 <h:outputText value="Kpi Name" />
                                             </f:facet>
-                                            <h:outputText value="#{model.kpiName}" />
-                                        </p:column>
-                                        <p:column>
-                                            <f:facet name="header">
-                                                <h:outputText value="Department Attached To" />
-                                            </f:facet>
-                                            <h:outputText value="#{model.department.name}" />
-                                        </p:column>
-                                        <p:column>
-                                            <f:facet name="header">
-                                                <h:outputText value="Goal Attached To" />
-                                            </f:facet>
-                                            <h:outputText value="#{model.goal.name}" />
+                                            <h:outputText value="#{model.title}" />
                                         </p:column>
                                         <p:column headerText="Description">
                                             <h:outputText value="#{model.description}" />
@@ -81,11 +69,11 @@
                                             <p:commandButton icon="fa fa-trash" title="Delete"
                                                              styleClass="ui-button-danger"
                                                              style="margin-left: 5px"
-                                                             actionListener="#{kpiView.deleteKpi()}"
-                                                             update="teamsView">
+                                                             actionListener="#{kpiView.deleteKpi}"
+                                                             update="kpisTable">
                                                 <f:setPropertyActionListener value="#{model}" target="#{kpiView.selectedKpi}" />
                                                 <p:confirm header="Confirmation"
-                                                           message="You are about to delete KPI '#{model.kpiName}'. Are you sure?"
+                                                           message="You are about to delete KPI '#{model.title}'. Are you sure?"
                                                            icon="ui-icon-alert" />
                                             </p:commandButton>
